fix: close source-reservation TOCTOU race and two stale-state crashes
Round 33 (Copilot PR #479 review):

- SourceInFlight/DocumentDeleter: an orphan-sweep "probe" (acquire-then-
  release) left a window between the probe and the delete for a new
  ingest to reserve the file. Replaced with acquire-and-hold: the
  reservation now spans the whole decision-and-removal window in both
  removeSourceFileIfUnreferenced() and removeSourceObjectUnderLock(),
  released only in a finally.
- SourceInFlight::reserve() now throws SourceReservationContendedException
  on contention instead of silently degrading to unprotected conversion;
  the new acquireForRemoval() keeps the deleting side's non-exceptional
  "keep it, it's in flight" behaviour. The contention exception is
  deliberately left uncaught by IngestDocumentJob and the Flow sync path
  so it fails the attempt loudly rather than converting unreserved.
- Preserved the artifacts-off/no-lock-store legacy behaviour (delete
  proceeds, protected only by the age grace) that the restructuring
  initially regressed twice during development; now covered by a
  dedicated regression test.

- DocumentIngestor::findExistingVersion() now queries withTrashed(): a
  soft-deleted row occupies the same tenant-scoped unique tuple the
  index protects, so an identical re-ingest after a soft delete used to
  miss the row and crash updateOrCreate() on the unique constraint. Both
  the idempotent short-circuit and the replaceExisting/forceReembed path
  through persistDocumentAndChunks() now restore a trashed match first
  (restoreIfTrashed(), with a metadata.restores provenance entry
  mirroring DocumentVersionService's shape).
- repairArtifactOfExistingVersion() now trusts document_hash, not the
  possibly-stale content_hash, as the expected hash — matching
  kb:artifacts-backfill's own choice — and recordContentHashIfDiffers()
  (renamed from recordContentHashIfMissing) corrects a present-but-wrong
  content_hash, not only a missing one.
- kb.ocr.job_timeout is no longer pre-cast in config: SourceInFlight
  validates it through SettingInt::whole(), and the earlier (int) cast
  would have truncated a fractional misconfiguration before that
  validator could refuse it. SettingIntTest now covers both this key and
  kb.sources.inflight_reservation_seconds.

Every fix has a dedicated regression test, each verified by reverting
the fix and confirming the test fails. Full suite: 4548 tests, 21055
assertions, green.

+++ config/kb.php
@@ -495,7 +495,15 @@ return [
         // the ingest job timeout and the re-run lock are all bounded by it —
         // and the queue's `retry_after` only has to exceed THIS (+ margins),
         // never a driver's theoretical worst case.
-        'job_timeout' => (int) env('KB_OCR_JOB_TIMEOUT', 3600),
+        //
+        // Kept RAW (no `(int)` cast, SEC-SETTING-SHAPE-001): `SourceInFlight
+        // ::defaultSeconds()` validates this through `SettingInt::whole()` —
+        // a cast here would truncate a fractional misconfiguration BEFORE
+        // that validator could refuse it, silently shortening the
+        // reservation a slow OCR run needs to outlive. `OcrService::
+        // runBudgetSeconds()` still casts its own local copy defensively;
+        // that consumer's behaviour is unchanged either way.
+        'job_timeout' => env('KB_OCR_JOB_TIMEOUT', 3600),
         'max_bytes' => (int) env('KB_OCR_MAX_BYTES', 26214400), // 25 MiB, the upload cap
         // Largest single figure a driver may hand back (remote drivers return
         // base64 images inside the response body).
